Break a dependency cycle.
